Replace maxent tagger with perceptron tagger
Before the changes in this commit, full analysis was taking roughly
500ms which feels too slow for an API. Using an ExecutorService
for populating the adjacencyList helped speed things up by about
10%, but that was still 450ms. I put a Cache in for vector
similarity calls and that reduced the time to as low as 5-6ms
which is wonderful in comparison. One thing to keep in mind would
be how much memory the server this runs on has. We'll need to set
a maximum size on the cache to prevent OOM. Also, we likely only
need the cache for the duration of a single article. Articles will
likely be redundant with themselves but there might not be as much
redundancy across articles.