Only use top N words instead of use a percentile
Using a map reduce approach means that percentiles are less valuable
and less accurate of an approach. The words are evaluated in smaller
subsets of the main corpus, so the words are less weighed against
their counter parts which produces worse results. I still want to
keep this in its own branch though in the case that I think of ways
to improve this approach.